Refactor: Add createCanvas util, Remove pixelRatio prop
-
