+++ packages/dogyeong/src/frontend/chart/index.ts
@@ -1,20 +1,8 @@
 import CandleChart from '@/chart/CandleChart';
+import { createCanvas } from '@/chart/utils';
 
 export const createChart = ($container: HTMLElement) => {
-  const canvas = document.createElement('canvas');
-  const width = $container.offsetWidth;
-  const height = $container.offsetHeight;
-  const pixelRatio = window.devicePixelRatio;
-  const ctx = canvas.getContext('2d');
+  const canvas = createCanvas($container);
 
-  canvas.style.width = `${width}px`;
-  canvas.style.height = `${height}px`;
-  canvas.width = width * pixelRatio;
-  canvas.height = height * pixelRatio;
-
-  ctx.scale(pixelRatio, pixelRatio);
-
-  $container.appendChild(canvas);
-
-  return new CandleChart(canvas, pixelRatio);
+  return new CandleChart(canvas);
 };
